+++ docs/CHANGELOG.md
@@ -7,12 +7,13 @@ Released: ???
 Features:
 - Added more Galatea fishing drops for potential future update to CT 3.0.
 - Renamed The Sea Emperor to The Loch Emperor everywhere.
+- Deleted outdated feature "Highlight matching items in Attribute Fusion" because there is no old attribute fusion anymore.
 
 Bugfixes:
 - 
 
 Other:
-- 
+-
 
 ## v1.45.0